+++ modules/core-config/modules/fluent-bit-aggregator/main.tf
@@ -1,9 +1,18 @@
+resource "kubectl_manifest" "resource_files" {
+  for_each = local.resource_files
+
+  yaml_body = file(each.value)
+
+  server_side_apply = true
+  wait              = true
+}
+
 resource "helm_release" "default" {
   name      = local.name
   namespace = var.namespace
 
-  repository = "https://grafana.github.io/helm-charts/"
-  chart      = "loki"
+  repository = "oci://ghcr.io/stevehipwell/helm-charts"
+  chart      = "fluent-bit-aggregator"
   version    = local.chart_version
   skip_crds  = true