+++ modules/core-config/modules/loki/identity.tf
@@ -7,7 +7,7 @@ module "identity" {
   workload_identity = local.use_aad_workload_identity
   oidc_issuer_url   = var.cluster_oidc_issuer_url
 
-  name      = "${var.cluster_name}-loki"
+  name      = "${var.cluster_name}-${local.name}"
   subjects  = ["system:serviceaccount:${var.namespace}:${local.service_account_name}"]
   namespace = var.namespace
   labels    = var.labels